moved main scripts to src folder for better organization
