+++ manifest.json
@@ -29,7 +29,7 @@
     }
   },
   "background": {
-    "service_worker": "service-worker.js",
+    "service_worker": "src/service-worker.js",
     "type": "module" 
   },
   "content_scripts": [
@@ -42,14 +42,14 @@
     }
   ],
   "side_panel": {
-    "default_path": "sidepanel.html"
+    "default_path": "src/sidepanel.html"
   },
   "web_accessible_resources": [
     {
       "resources": [
-        "legalTerms.js", 
+        "src/legalTerms.js", 
         "sidepanel.html", 
-        "styles.css",
+        "src/styles.css",
         "node_modules/cheerio/lib/cheerio.min.js",
         "node_modules/compromise/builds/compromise.min.js"
       ],
