+++ src/legalTerms.js
