+++ src/modelHandler.js
@@ -1,0 +1,70 @@
+/**
+ * @file modelHandler.js
+ * @description This file contains the ModelHandler class that handles loading a model and analyzing text using the model.
+ * @version 1.0.0
+ * @date 2024-09-25
+ * 
+ * @author Timmothy Escolopio
+ * @company 3D Tech Solutions LLC
+ * 
+ * @changes
+ *  - 2024-09-25 | tescolopio | Initial creation of the script.
+ */
+
+import * as tf from '@tensorflow/tfjs';
+
+export class ModelHandler {
+    constructor() {
+        this.model = null;
+    }
+
+    async loadModel(modelUrl) {
+        try {
+            this.model = await tf.loadLayersModel(modelUrl);
+            console.log('Model loaded successfully');
+        } catch (error) {
+            console.error('Error loading model:', error.message, error.stack);
+        }
+    }
+
+    async analyzeText(text) {
+        if (!this.model) {
+            console.error('Model not loaded');
+            return 'Error: Model not loaded.';
+        }
+
+        try {
+            // Preprocess text for model input here. The preprocessing will depend on the model's expected input format.
+            const processedText = this.preprocessText(text);
+
+            // Convert processed text to tensor or appropriate format for the model input
+            const inputTensor = tf.tensor2d([processedText], [1, processedText.length]); // Adjust shape according to your model's input
+
+            // Perform prediction
+            const prediction = this.model.predict(inputTensor);
+
+            // Postprocess model output to get the content grade. This might involve converting tensor outputs to human-readable format or grades.
+            const contentGrade = await this.postprocessPrediction(prediction);
+
+            return contentGrade;
+        } catch (error) {
+            console.error('Error analyzing text:', error.message, error.stack);
+            return 'Error: Problem during the analysis.';
+        }
+    }
+
+
+    async postprocessPrediction(prediction) {
+        // Placeholder for postprocessing logic
+        // This should convert the model's prediction to a grade or other meaningful output
+        try {
+            const predictionArray = await prediction.array();
+            console.log('Prediction output:', predictionArray);
+            // This is a simplification. Replace it with actual logic to interpret your model's output.
+            return 'A'; // Example grade, replace with actual postprocessing to interpret the model's output
+        } catch (error) {
+            console.error('Error processing prediction output:', error.message, error.stack);
+            return 'Error: Problem during postprocessing.';
+        }
+    }
+}
