+++ src/rightsAssessor.js
