+++ src/sidepanel.html
