+++ src/sidepanel.js
