+++ src/summarizeTos.js
