+++ src/textExtractor.js
